Moved unshared state down to IngredientForm components
