+++ src/components/forms/editRecipe/formUI/EditIngredientForm.tsx
@@ -1,21 +1,17 @@
 'use client'
 import { AttachedIngredient, Ingredient } from '@/types/ingredientType';
 import { validateQuantityInput } from '@/utils/helpers/validateQuantityInput';
-import { Dispatch, KeyboardEvent, MouseEvent, SetStateAction } from 'react';
+import { Dispatch, KeyboardEvent, MouseEvent, SetStateAction, useState } from 'react';
 import Select from 'react-select';
 import styles from "../../recipeForm.module.css"
 
 interface EditIngredientFormProps {
     initialIngredients: AttachedIngredient[],
     allIngredients: Ingredient[],
-    markedForDeletion: (ingredientObject: AttachedIngredient) => boolean,
-    handleUndoDelete: (event: MouseEvent<HTMLButtonElement>, objectToUndo: AttachedIngredient) => void,
-    handleDeleteExistingIngredient: (event: MouseEvent<HTMLButtonElement>, objectToDelete: AttachedIngredient) => void,
-    handleRemoveIngredient: (event: MouseEvent<HTMLButtonElement>, objectToRemove: AttachedIngredient) => void,
     ingredientsToPost: AttachedIngredient[],
-    ingredientToAdd: AttachedIngredient,
-    updateIngredientToAdd: Dispatch<SetStateAction<AttachedIngredient>>,
-    handleAddIngredient: (event: MouseEvent<HTMLButtonElement> | KeyboardEvent<HTMLInputElement>) => void,
+    ingredientsToDelete: AttachedIngredient[],
+    updateIngredientsToPost: Dispatch<SetStateAction<AttachedIngredient[]>>,
+    updateIngredientsToDelete: Dispatch<SetStateAction<AttachedIngredient[]>>,
     setShowCustom: Dispatch<SetStateAction<boolean>>
 }
 
@@ -28,9 +24,102 @@ interface EditIngredientFormProps {
  * Newly added ingredients are presented to the user with a green highlight, deleted ingredients with a red highlight,
  * and unchanged ingredients remain neutral. 
   */
-export const EditIngredientForm = ({ initialIngredients, allIngredients, markedForDeletion, handleUndoDelete,
-    handleDeleteExistingIngredient, handleRemoveIngredient, ingredientsToPost, ingredientToAdd,
-    updateIngredientToAdd, handleAddIngredient, setShowCustom }: EditIngredientFormProps) => {
+export const EditIngredientForm = ({ initialIngredients, allIngredients, ingredientsToPost, ingredientsToDelete, 
+    updateIngredientsToPost, updateIngredientsToDelete, setShowCustom }: EditIngredientFormProps) => {
+
+    const [ingredientToAdd, updateIngredientToAdd] = useState<AttachedIngredient>(
+        {
+            "ingredient": 0,
+            "name": "",
+            "quantity": "",
+            "quantity_unit": ""
+        }
+    )
+
+    const handleAddIngredient = (event: MouseEvent<HTMLButtonElement> | KeyboardEvent<HTMLInputElement>) => {
+        event.preventDefault()
+        // Check if required fields have been entered
+        if (ingredientToAdd.ingredient > 0 && ingredientToAdd.quantity.length > 0) {
+            // Get a copy of the ingredientToAdd
+            const copyIngToAdd = { ...ingredientToAdd }
+            // Trim any whitespace on the quantity
+            copyIngToAdd.quantity = ingredientToAdd.quantity.trim()
+            updateIngredientToAdd(copyIngToAdd)
+
+            // Get a copy of the current array of ingredients that are staged to be added
+            const copy = [...ingredientsToPost]
+            // Check if the ingredient has already been added
+            const alreadyAdded = copy.some(ingredient => ingredient.ingredient === copyIngToAdd.ingredient)
+            const stagedFordeletion = ingredientsToDelete.some(ingredient => ingredient.ingredient === copyIngToAdd.ingredient)
+            const inInitialRecipe = initialIngredients.some(ingredient => ingredient.ingredient === copyIngToAdd.ingredient)
+            if (!alreadyAdded && !inInitialRecipe) { // If not already staged to post and not in initial recipe, add to array to post
+                copy.push(copyIngToAdd)
+                updateIngredientsToPost(copy)
+            } else if (inInitialRecipe && stagedFordeletion && !alreadyAdded) { // If in initial recipe but staged for deletion, allow user to add new ingredient to edit quantity/measure
+                copy.push(copyIngToAdd)
+                updateIngredientsToPost(copy)
+            } else {
+                window.alert("That ingredient has already been added")
+            }
+        } else {
+            window.alert("Please enter an ingredient and a quantity")
+        }
+    }
+
+    /**Removes a recently added ingredient that was not originally on recipe */
+    const handleRemoveIngredient = (event: MouseEvent<HTMLButtonElement>, objectToRemove: AttachedIngredient) => {
+        event.preventDefault()
+        const updatedIngredients = ingredientsToPost.filter(ingredient => ingredient.ingredient !== objectToRemove.ingredient)
+        updateIngredientsToPost(updatedIngredients)
+    }
+
+    /**Stages ingredient originally on recipe for deletion */
+    const handleDeleteExistingIngredient = (event: MouseEvent<HTMLButtonElement>, objectToDelete: AttachedIngredient) => {
+        event.preventDefault()
+
+        // Get a copy of the current array of ingredients that are staged to be deleted
+        const copy = [...ingredientsToDelete]
+
+        // Check if the ingredient has already been staged for deletion
+        const alreadyStaged = copy.some(ingredient => ingredient.ingredient === objectToDelete.ingredient)
+
+        if (!alreadyStaged) {
+            copy.push(objectToDelete)
+            updateIngredientsToDelete(copy)
+        } else {
+            window.alert("Ingredient already marked for deletion")
+        }
+
+    }
+
+    /**Unstages ingredient that was originally on recipe and has been staged for deletion */
+    const handleUndoDelete = (event: MouseEvent<HTMLButtonElement>, objectToUndo: AttachedIngredient) => {
+        event.preventDefault()
+        // Get a copy of the current array of ingredients that are staged to be added and deleted
+        const addedIngCopy = [...ingredientsToPost]
+        const IngToDeleteCopy = [...ingredientsToDelete]
+        // Check if the same ingredient has been added since being staged for deletion
+        const changesMadeToIngredient = addedIngCopy.some(ingredient => ingredient.ingredient === objectToUndo.ingredient)
+
+        if (changesMadeToIngredient) { // If changes for this ingredient have been added
+            // Remove the new ingredient object with the changes
+            const updatedIngredientsToPost = addedIngCopy.filter(ingredient => ingredient.ingredient !== objectToUndo.ingredient)
+            updateIngredientsToPost(updatedIngredientsToPost)
+            // Then, remove the ingredient object from the ingredients array staged for deletion
+            const updatedIngredientsToRemove = IngToDeleteCopy.filter(ingredient => ingredient.ingredient !== objectToUndo.ingredient)
+            updateIngredientsToDelete(updatedIngredientsToRemove)
+        } else { // else, only remove the ingredient object from the ingredients array staged for deletion
+            const updatedIngredientsToRemove = IngToDeleteCopy.filter(ingredient => ingredient.ingredient !== objectToUndo.ingredient)
+            updateIngredientsToDelete(updatedIngredientsToRemove)
+        }
+
+    }
+
+    /**Checks if an ingredient has been marked for deletion. Returns true or false */
+    const markedForDeletion = (ingredientObject: AttachedIngredient) => {
+        const alreadyStaged = ingredientsToDelete.some(ingredient => ingredient.ingredient === ingredientObject.ingredient)
+        return alreadyStaged
+    }
 
     return <>
         <div className={styles["addedIngredients"]}>
@@ -38,7 +127,7 @@ export const EditIngredientForm = ({ initialIngredients, allIngredients, markedF
                 initialIngredients.length > 0
                 && initialIngredients.map(initialIngredient => {
                     return <div key={`displayInitIng--${initialIngredient.ingredient}`}>
-                        { // If user has marked for delete, show 'undo' button
+                        { // If user has marked for delete, display with red highlight and show 'undo' button
                             markedForDeletion(initialIngredient)
                                 ?
                                 <div className={`${styles["addedIngredientRow"]} ${styles["ingredientToDelete"]}`} key={`initialIngDetails1--${initialIngredient.ingredient}`}>
@@ -54,7 +143,7 @@ export const EditIngredientForm = ({ initialIngredients, allIngredients, markedF
                                     </span>
                                 </div>
 
-                                :
+                                : /*Else, display with neutral color and delete button */
                                 <div className={styles["addedIngredientRow"]} key={`initialIngDetails2--${initialIngredient.ingredient}`}>
                                     <span className={styles["flex-column1"]}>{initialIngredient.name}</span>
                                     <span className={styles["flex-column2"]}>{initialIngredient.quantity} {initialIngredient.quantity_unit}</span>
@@ -68,13 +157,12 @@ export const EditIngredientForm = ({ initialIngredients, allIngredients, markedF
                                             className={styles["btn--removeItem"]}>X</button>
                                     </span>
                                 </div>
-
                         }
                     </div>
                 })
             }
 
-            {   /*New ingredients being added to recipe card by user*/
+            {   /*New ingredients being added to recipe card by user, displayed with green highlight*/
                 ingredientsToPost.length > 0
                 && ingredientsToPost.map(includedIngredient => {
                     return <div className={`${styles["addedIngredientRow"]} ${styles["addedIngredientRow--editForm"]}`} key={`addedIngDetails--${includedIngredient.ingredient}`}>
